refactor(svelte): split legend focus pure unit test monolith
Move shared SceneLegend fixtures to focus-fixtures and split the 1161-line
focus.test.ts by production module: focus helpers, pure entry-key-index,
focus-emphasis, and focus-plans. Assertions preserved.
